+++ prometheus/ready_test.go
@@ -2,12 +2,12 @@ package prometheus
 
 import (
 	"context"
-	"net/http"
-	"net/http/httptest"
 	"net/url"
 	"testing"
+	"time"
 
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/config"
+	"github.com/lightstep/opentelemetry-prometheus-sidecar/internal/promtest"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/telemetry"
 	"github.com/stretchr/testify/require"
 )
@@ -15,30 +15,39 @@ import (
 var logger = telemetry.DefaultLogger()
 
 func TestReady(t *testing.T) {
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusOK)
-	}))
+	fs := promtest.NewFakePrometheus()
+	fs.SetReady(true)
+	fs.SetIntervals(30 * time.Second)
 
-	tu, err := url.Parse(ts.URL)
-	require.NoError(t, err)
+	require.NoError(t, WaitForReady(context.Background(), fs.ReadyConfig()))
+}
 
-	require.NoError(t, WaitForReady(context.Background(), logger, tu))
+func TestSlowStart(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping test in short mode.")
+	}
+	fs := promtest.NewFakePrometheus()
+	fs.SetReady(true)
+	fs.SetIntervals()
+
+	go func() {
+		time.Sleep(time.Second * 3)
+		fs.SetIntervals(30)
+	}()
+
+	require.NoError(t, WaitForReady(context.Background(), fs.ReadyConfig()))
 }
 
 func TestNotReady(t *testing.T) {
 	if testing.Short() {
 		t.Skip("skipping test in short mode.")
 	}
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusBadRequest)
-	}))
-
-	tu, err := url.Parse(ts.URL)
-	require.NoError(t, err)
+	fs := promtest.NewFakePrometheus()
+	fs.SetReady(false)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 2*config.DefaultHealthCheckTimeout)
 	defer cancel()
-	err = WaitForReady(ctx, logger, tu)
+	err := WaitForReady(ctx, fs.ReadyConfig())
 	require.Error(t, err)
 	require.Equal(t, context.DeadlineExceeded, err)
 }
@@ -53,21 +62,69 @@ func TestReadyFail(t *testing.T) {
 
 	ctx, cancel := context.WithTimeout(context.Background(), 2*config.DefaultHealthCheckTimeout)
 	defer cancel()
-	err = WaitForReady(ctx, logger, tu)
+	err = WaitForReady(ctx, config.PromReady{
+		Logger:  logger,
+		PromURL: tu,
+	})
 	require.Error(t, err)
 }
 
 func TestReadyCancel(t *testing.T) {
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusOK)
-	}))
-
-	tu, err := url.Parse(ts.URL)
-	require.NoError(t, err)
+	fs := promtest.NewFakePrometheus()
 
 	ctx, cancel := context.WithCancel(context.Background())
 	cancel() // immediate
-	err = WaitForReady(ctx, logger, tu)
+	err := WaitForReady(ctx, fs.ReadyConfig())
+
 	require.Error(t, err)
 	require.Equal(t, context.Canceled, err)
+}
+
+func TestReadySpecificInterval(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping test in short mode.")
+	}
+	fs := promtest.NewFakePrometheus()
+	fs.SetIntervals() // None set
+
+	const interval = 79 * time.Second
+
+	rc := fs.ReadyConfig()
+	rc.ScrapeIntervals = []time.Duration{time.Minute, interval}
+
+	go func() {
+		time.Sleep(1 * time.Second)
+		fs.SetIntervals(20 * time.Second)
+
+		time.Sleep(1 * time.Second)
+		fs.SetIntervals(20*time.Second, 40*time.Second)
+
+		time.Sleep(1 * time.Second)
+		fs.SetIntervals(20*time.Second, 40*time.Second, 60*time.Second)
+
+		time.Sleep(1 * time.Second)
+		fs.SetIntervals(20*time.Second, 40*time.Second, 60*time.Second, interval)
+	}()
+
+	err := WaitForReady(context.Background(), rc)
+
+	require.NoError(t, err)
+}
+
+func TestReadySpecificIntervalWait(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping test in short mode.")
+	}
+	fs := promtest.NewFakePrometheus()
+	fs.SetIntervals(19 * time.Second) // Not the one we want
+
+	rc := fs.ReadyConfig()
+	rc.ScrapeIntervals = []time.Duration{79 * time.Second, 19 * time.Second}
+
+	ctx, cancel := context.WithTimeout(context.Background(), config.DefaultHealthCheckTimeout*4/3)
+	defer cancel()
+	err := WaitForReady(ctx, rc)
+
+	require.Error(t, err)
+	require.Equal(t, context.DeadlineExceeded, err)
 }